HIVE-2894 [jira] RCFile Reader doesn't provide access to Metadata (Owen O'Malley via Ashutosh Chauhan)
Summary:
hive-2894

Add an accessor for RCFile's metadata.

Currently the RCFile writer can add metadata to an RCFile, but the reader
doesn't provide an accessor. I'd like to add one.

Test Plan:
I added a call to test that the metadata that was passed in was available from
the reader.

Reviewers: JIRA, ashutoshc

Reviewed By: ashutoshc

Differential Revision: https://reviews.facebook.net/D2421

git-svn-id: https://svn.apache.org/repos/asf/hive/trunk@1304693 13f79535-47bb-0310-9956-ffa450edef68