HIVE-23941: Refactor TypeCheckProcFactory to be database agnostic (Steve Carlin, reviewed by Jesus Camacho Rodriguez)
Part of the code has already been refactored to become database agnostic
(i.e. HiveFunctionHelper).

Further refactoring needs to be done on TypeCheckProcFactory which also
should be database agnostic.

The most interesting part of this code change involved the addition of
HiveFunctionInfo.  Previous to this commit, a query specific GenericUDF was
cloned out of the FunctionRegistry every time that the FunctionInfo.getGenericUDF()
was called. With this commit, a new query specific FunctionInfo (HiveFunctionInfo)
is created with the cloned GenericUDF.

Closes apache/hive#1326