HIVE-25625: Drop DummyCliDriver, TestCompareCliDriver from tests (Stamatis Zampetakis, reviewed by Zoltan Haindrich)
1. Drop DummyCliDriver

The only thing this test code does is fail no matter the input file
(potentially with different message). It is very close to "dead-code" so
keeping it in the repo only adds maintenance overhead.

2. Drop TestCompareCliDriver

The driver has been introduced back in 2015 aiming to run queries with
vectorization on/off and comparing the results. However it didn't
receive much attention since then and currently only two queries are run
with this driver. The majority of tests aiming to ensure vectorization
works correctly use the TestMiniLlapLocalCliDriver and run a query
twice switching on/off the necessary properties.

Summing up having the CoreCompareCliDriver in the repo leads to extra
code maintenance cost without much benefit.

3. Remove unused classes/methods after dropping Compare/Dummy drivers

Closes #2732