HIVE-26437: Dump unpartitioned tables in parallel. (Amit Saonerkar reviewed by Teddy Choi, Rajesh Balamohan)
* HIVE-26437: Dump unpartitioned tables in parallel.

Currently partitions of table is dump in parallel manner. But if table is not partitioned, it is dumped serially.
This change introduces parallelism at table level as well. A single thread pool which is currently being used
for partition level is also used for table level. The table level dump task is added to same thread pool.
The degree of parallelism depends upon config parameter REPL_PARTITIONS_DUMP_PARALLELISM whose defaul value is 100.
The new ExportService is introduced with this change which would be responsible for exporting table and partitions during repl dump. The ExportService is initialized and configured with thread pools by HiveServer2 service.
A new Hiveconfig variable ie "REPL_TABLE_DUMP_PARALLELISM is introduced to define the number of threads which would be created in thread pool.
The ExportService which is created as singleton instance would be used by ReplDumpTask.

Change-Id: I16839e30a51e886da6eb707034e1c44bcd0a13dc

* Make ExportService member variables final and volatile to have single copy across each threads.
Qualify ExportService methods as synchronized
Make sure SessionState is accessed only when it is non null.
Create and Configure ExportService inside HiveServer2.
SessionState configures ExportService only if it is found to be not running.

Change-Id: I437825b1360fb8bc16537f85a2929cd07c999361

* Changed design of ExportService. Created it as non shareable instance.
ExportService instance is now created when ReplDumptask gets executed.
The service is shutdown as soon as table and partition dump process is over.

Change-Id: I13921c76832cfcd8c6a86f6b042f73fb7521561f

* Removed a testcase from TestExportService due to the newly added function in ExportService ie waitForTaskToFinishAndShutdown().

Change config REPL_TABLE_DUMP_PARALLELISM default value to 15.

Change-Id: Id87bd2458cf9ee5759866eefd51feab0f00944a4

* Make ExportService instance common to all database which are being dumped.

Change-Id: I60bdcc628af372cacfa00896dd9307c871936686

* Re-trigger CI

Change-Id: Ibef10d8ae4188a92dab8ffdc241d0cf212b513f5