HIVE-26683: Sum windowing function returns wrong value when all nulls. (#3800)
* HIVE-26683: Sum windowing function returns wrong value when all nulls.

The sum windowing function is returning an incorrect value when all the
"following" rows are null.  The correct value for sum when all the rows
are null is "null".

A new member variable had to be added to track for nulls. It uses the
same algorithm that is used for sums. The sums are tracked by keeping
a running sum across all the rows and subtracting off the running sum
outside the window. Likewise, we keep track of a running non null row
count for the current row and subtract the non null row count of the
row that is leaving the window.

* empty