HIVE-27217: Rethrow unhandled exception in addWriteNotificationLogInBatch (John Sherman reviewed by Stephen Carlin, Stamatis Zampetakis)
It is not a good choice to swallow the exception. A missing write
notification will affect various systems such as event notification and
replication. It is much preferable to fail and make it visible so a
user can take action (either by checking HMS logs, making sure HMS
connectivity is good, etc) rather than having to figure out why the
notification log is missing entries long after the fact.

This is likely an oversight when attempting the fallback mechanism
rather than a purposeful choice to ignore these exceptions.

Closes #4197