HIVE-27504: Automate inclusion of 3-party licenses in binary distribution (#4484) (Stamatis Zampetakis, Reviewed by Attila Turoczy, Sai Hemanth Gantasala, Ayush Saxena)
* HIVE-27504: Automate inclusion of 3-party licenses in binary distribution

Manually downloading and including licenses of bundled dependencies in
the binary distribution is errorprone and time consuming. Moreover,
dependencies and their versions change rapidly and it is hard to keep
the LICENSE documentation correct and up-to-date.

In this change, we make use of license and assembly plugin to automate
this process and remove the need for committing every license in git.
In addition, this alleviates the need for regular updates and extra
commits when the respective dependencies change (add/remove/version
bump).

Brief outline of the changes:
1. Add and configure maven-license-plugin for downloading licenses
for bundled dependencies.
2. Remove manual curated binary-package-licenses directory.
3. Configure assembly plugin to included downloaded licenses in the
binary distribution.
4. Use maven-ant-plugin to generated the root LICENSE file for the
binary distro.

This change adds automation only as part of releasing Hive as a whole.
Releasing metastore and storage api requires more changes in the
respective modules but this is out of scope for now.

As part of the automation this change fixes various issues in the content
structure of the LICENSE and NOTICE file.

1. LICENSE/NOTICE should reflect exactly what lies inside each distro
(source & binary); ideally no more and definitely not less.
2. LICENSES for bundled dependencies should be referenced from the main
LICENSE file not from NOTICE (or other places) as it was before this
change.
3. For bundled dependencies the versions should be mentioned in the
LICENSE file cause different version may use different licenses and it
should be clear which version is bundled.

* Format xml files with xmlstarlet