HIVE-26818: Beeline module misses transitive dependencies due to shading (Stamatis Zampetakis reviewed by Attila Turoczy, Ayush Saxena, Kokila N)
Due to shading, the dependecy-reduced-pom.xml file is installed in the local maven repository (~/.m2/repository/org/apache/hive/hive-beeline/4.0.0-SNAPSHOT/) for beeline. The latter indicates that the module doesn't have any transitive dependencies. If we were publishing the shaded jar that would be true but we publish the regular jar. At this point, modules which include hive-beeline as a maven dependency are broken leading to build problems and classpath errors.

1. Do not generate the dependency-reduced-pom.xml and publish the regular pom.xml file along with the main jar file.
2. Remove some redundant explicit dependency declarations that were added to workaround the missing dependency problem from the beeline module.

Close apache/hive#5055